time: move error types into `time::error
