+++ tokio/src/time/timeout.rs
@@ -4,10 +4,9 @@
 //!
 //! [`Timeout`]: struct@Timeout
 
-use crate::time::{sleep_until, Duration, Instant, Sleep};
+use crate::time::{sleep_until, Duration, Elapsed, Instant, Sleep};
 
 use pin_project_lite::pin_project;
-use std::fmt;
 use std::future::Future;
 use std::pin::Pin;
 use std::task::{self, Poll};
@@ -112,18 +111,6 @@ pin_project! {
     }
 }
 
-/// Error returned by `Timeout`.
-#[derive(Debug, PartialEq)]
-pub struct Elapsed(());
-
-impl Elapsed {
-    // Used on StreamExt::timeout
-    #[allow(unused)]
-    pub(crate) fn new() -> Self {
-        Elapsed(())
-    }
-}
-
 impl<T> Timeout<T> {
     pub(crate) fn new_with_delay(value: T, delay: Sleep) -> Timeout<T> {
         Timeout { value, delay }
@@ -161,24 +148,8 @@ where
 
         // Now check the timer
         match me.delay.poll(cx) {
-            Poll::Ready(()) => Poll::Ready(Err(Elapsed(()))),
+            Poll::Ready(()) => Poll::Ready(Err(Elapsed::new())),
             Poll::Pending => Poll::Pending,
         }
-    }
-}
-
-// ===== impl Elapsed =====
-
-impl fmt::Display for Elapsed {
-    fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
-        "deadline has elapsed".fmt(fmt)
-    }
-}
-
-impl std::error::Error for Elapsed {}
-
-impl From<Elapsed> for std::io::Error {
-    fn from(_err: Elapsed) -> std::io::Error {
-        std::io::ErrorKind::TimedOut.into()
     }
 }
